refactor(hickey+lowy): extract wire-link internals — framing codec + retry policy
F4 (hickey): move the base64+newline framing codec (encodeFrame/decodeFrame/
readFramedLines) out of links/stdio.ts into an unexported links/stdio-codec.ts,
shared with peer-server.ts — one home for the framing change-axis, off the
public link export surface.
D (lowy cross-validation): the ClientRetryPlugin install was duplicated across
websocketLink + stdioLink; extract links/_wire.ts (wireRetryPlugins + wireClient)
so the retry policy has one home and a future wire link inherits it for free.
